requirements.txt
certifi==2020.6.20
gitdb==4.0.5
GitPython==3.1.11
Jinja2==2.11.2
MarkupSafe==1.1.1
smmap==3.0.4
